% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sentinel2.R
\name{addSen2queue}
\alias{addSen2queue}
\title{Copy the scenes over area of extent and time span of interest from an input folder to a level 1 folder and add these scenes to a queue file}
\usage{
addSen2queue(
  S2auxfolder,
  ext,
  S2folder,
  starttime,
  endtime,
  queuefolder,
  queuefile,
  l1folder
)
}
\arguments{
\item{S2auxfolder}{directory with auxilliarity Sentinel-2 data, e.g. tile grid}

\item{ext}{extent of the area of interest (vector with xmin, xmax, ymin, ymax in degrees)}

\item{S2folder}{directory where all Sentinel-2 data are stored}

\item{starttime}{start date of the time span of interest (vector with year, month, day)}

\item{endtime}{end date of the time span of interest (vector with year, month, day)}

\item{queuefolder}{directory where the queue file is stored}

\item{queuefile}{name of the queue file}

\item{l1folder}{directory of the level 1 folder}
}
\value{
adds Sentinel-2 data to the level1 folder and queue
}
\description{
Copy the scenes over area of extent and time span of interest from an input folder to a level 1 folder and add these scenes to a queue file
}
