% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{cfg_to_list}
\alias{cfg_to_list}
\title{Convert config data frame to list}
\usage{
cfg_to_list(cfg)
}
\arguments{
\item{cfg}{The configuration data frame}
}
\value{
The configuration list
}
\description{
Using a list instead of a data frame can be practical. For instance, it
allows using the $ operator for extracting values.
}
