% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{cfgl_to_df}
\alias{cfgl_to_df}
\title{Convert config list to data frame}
\usage{
cfgl_to_df(cfgl)
}
\arguments{
\item{cfgl}{The configuration list}
}
\value{
The configuration data frame
}
\description{
Convert config list to data frame
}
