% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{checkLSlog}
\alias{checkLSlog}
\title{Screenes the FORCE log file and add scenes to logfiles dependent on their processing category. Four categories are defined:
successful processing, failed due to unrecognized mission, skipped, and other}
\usage{
checkLSlog(
  scenes,
  logfolder,
  Sskiplogfile,
  Ssuccesslogfile,
  Smissionlogfile,
  Sotherlogfile
)
}
\arguments{
\item{scenes}{the names of the log files to be screened}

\item{logfolder}{full path to the folder where the logfiles are located}

\item{Sskiplogfile}{full path to the log file for skipped scenes}

\item{Ssuccesslogfile}{full path to the log file for successful scenes}

\item{Smissionlogfile}{full path to the log file for scenes with unknown mission}

\item{Sotherlogfile}{full path to the log file for other scenes}
}
\value{
adds scenes to logfile based on processing category
}
\description{
Screenes the FORCE log file and add scenes to logfiles dependent on their processing category. Four categories are defined:
successful processing, failed due to unrecognized mission, skipped, and other
}
