% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{createFireStack}
\alias{createFireStack}
\title{Convert CCI fire stack to a stack with a predefined temporal resolution and containing the value 0 when no fire is present and 1 if a fire is present}
\usage{
createFireStack(x, dts, resol, thres)
}
\arguments{
\item{x}{stack of CCI fire imagery, the first layer contains a mask (here the value 0 denotes that the pixel should not be included, the value 1 denotes that the pixel should be included), followed by n fire confidence layers and n fire doy layers}

\item{dts}{dates associated with the stack}

\item{resol}{the desired temporal resolution of the output data}

\item{thres}{threshold on the fire confidence layer, only fires having a confidence higher than the threshold are included}
}
\value{
a stack with with a predefined temporal resolution and containing the value 0 when no fire is present and 1 if a fire is present
}
\description{
Convert CCI fire stack to a stack with a predefined temporal resolution and containing the value 0 when no fire is present and 1 if a fire is present
}
