% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetools.R
\name{date_to_vec}
\alias{date_to_vec}
\title{Convert date to vector}
\usage{
date_to_vec(date)
}
\arguments{
\item{date}{A date object}
}
\value{
A vector in the format c(2021, 1, 16)
}
\description{
This is an auxiliary function.
}
