% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTM_DEM.R
\name{dllDEM}
\alias{dllDEM}
\title{Download SRTM DEM at 1 arcsecond
https://lpdaac.usgs.gov/products/srtmgl1v003/}
\usage{
dllDEM(ext, dl_dir = Sys.getenv("HOME"))
}
\arguments{
\item{ext}{vector with xmin, xmax, ymin, ymax in degrees}

\item{dl_dir}{where do you want to store the data?}
}
\value{
stores files to disk
}
\description{
Download SRTM DEM at 1 arcsecond
https://lpdaac.usgs.gov/products/srtmgl1v003/
}
