% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTM_DEM.R
\name{dllLPDAAC}
\alias{dllLPDAAC}
\title{Download data from LP DAAC DATa Pool}
\usage{
dllLPDAAC(dl_dir = Sys.getenv("HOME"), files)
}
\arguments{
\item{dl_dir}{where do you want to store the data?}

\item{files}{url of the files to download. This should either be (i) a single url, (ii) a list of urls, or (iii) the full path the a text file that contains the urls of interest}
}
\value{
Downloads data from LP DAAC Data Pool
}
\description{
Download data from LP DAAC DATa Pool
}
