% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landsat.R
\name{dllLS}
\alias{dllLS}
\title{Download Landsat scenes. Several criteria can be set to select scenes of interest: the extent of the study area,
the time period of interest, the preferred sensors, tiers and maximum cloud percentage.
Only scenes are downloaded that are not in the queue file yet. After download to a temporary folder, these scenes
 are compressed to a .tar.gz file, added to the the queue file and the level1 folder.
 Finally, the original downloaded files are removed from the temporary folder.}
\usage{
dllLS(
  l1folder,
  queuefolder,
  queuefile,
  tmpfolder,
  logfile,
  ext,
  starttime,
  endtime,
  sensors,
  tiers,
  cld
)
}
\arguments{
\item{l1folder}{level 1 folder}

\item{queuefolder}{queue folder}

\item{queuefile}{queue file}

\item{tmpfolder}{folder with temporary data}

\item{ext}{extent of the study area of interest, this should be given as a vector containing - in order - xmin, xmax, ymin, ymax in degrees}

\item{starttime}{start date (given as a vector with year, month, day)}

\item{endtime}{end date (given as a vector with year, month, day)}

\item{sensors}{sensors of interest, can be LC08, LE07, LT05, LT04, LM05, LM04}

\item{tiers}{tiers of interest, T1 and T2 represent Tier 1 and 2, respectively}

\item{cld}{maximum cloud percentage}
}
\value{
adds Landsat data to Level 1 folder
}
\description{
Download Landsat scenes. Several criteria can be set to select scenes of interest: the extent of the study area,
the time period of interest, the preferred sensors, tiers and maximum cloud percentage.
Only scenes are downloaded that are not in the queue file yet. After download to a temporary folder, these scenes
 are compressed to a .tar.gz file, added to the the queue file and the level1 folder.
 Finally, the original downloaded files are removed from the temporary folder.
}
