% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTM_DEM.R
\name{EartDataLogin}
\alias{EartDataLogin}
\title{Set Up Direct Access the LP DAAC Data Pool with R
The function configures a netrc profile that will allow users to download data from
 an Earthdata Login enabled server.}
\usage{
EartDataLogin()
}
\value{
generates a .netrc profile
}
\description{
Set Up Direct Access the LP DAAC Data Pool with R
The function configures a netrc profile that will allow users to download data from
 an Earthdata Login enabled server.
}
