% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{export_params}
\alias{export_params}
\title{Export parameters to file}
\usage{
export_params(config, file)
}
\arguments{
\item{config}{Configuration data frame (as loaded by parse_params)}

\item{file}{Filename where the parameters will be exported}
}
\value{
Nothing
}
\description{
Export parameters to file
}
