% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{ext_overlap}
\alias{ext_overlap}
\title{Check if a polygon and raster contain each other, intersect, or are not overlapping}
\usage{
ext_overlap(pol, rst)
}
\arguments{
\item{pol}{polygon}

\item{rst}{SpatRaster object}
}
\value{
1 if raster contains the polygon, 2 if polygon contains the raster or polygon, 3 if polygon and raster intersect,
and 3 if there is no overlap
}
\description{
Check if a polygon and raster contain each other, intersect, or are not overlapping
}
