% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{gen_params}
\alias{gen_params}
\title{Auxiliary function for generating a parameters' data frame}
\usage{
gen_params(
  FILE_QUEUE = "data/level1/queue.txt",
  DIR_LEVEL2 = "data/level2",
  DIR_LOG = "data/log",
  DIR_TEMP = "data/temp",
  FILE_DEM = "data/misc/dem/srtm.vrt",
  DEM_NODATA = "-32768",
  DO_REPROJ = "TRUE",
  DO_TILE = "TRUE",
  FILE_TILE = "NULL",
  TILE_SIZE = "3000",
  BLOCK_SIZE = "300",
  RESOLUTION_LANDSAT = "30",
  RESOLUTION_SENTINEL2 = "10",
  ORIGIN_LON = "-90",
  ORIGIN_LAT = "60",
  PROJECTION = "GLANCE7",
  RESAMPLING = "NN",
  DO_ATMO = "TRUE",
  DO_TOPO = "TRUE",
  DO_BRDF = "TRUE",
  ADJACENCY_EFFECT = "TRUE",
  MULTI_SCATTERING = "TRUE",
  DIR_WVPLUT = "data/misc/wvp",
  WATER_VAPOR = "NULL",
  DO_AOD = "TRUE",
  DIR_AOD = "NULL",
  ERASE_CLOUDS = "FALSE",
  MAX_CLOUD_COVER_FRAME = "75",
  MAX_CLOUD_COVER_TILE = "75",
  CLOUD_BUFFER = "300",
  SHADOW_BUFFER = "90",
  SNOW_BUFFER = "30",
  CLOUD_THRESHOLD = "0.225",
  SHADOW_THRESHOLD = "0.02",
  RES_MERGE = "REGRESSION",
  DIR_COREG_BASE = "NULL",
  COREG_BASE_NODATA = "-9999",
  IMPULSE_NOISE = "TRUE",
  BUFFER_NODATA = "FALSE",
  TIER = "1",
  NPROC = "1",
  NTHREAD = "1",
  PARALLEL_READS = "FALSE",
  DELAY = "10",
  TIMEOUT_ZIP = "30",
  OUTPUT_FORMAT = "GTiff",
  OUTPUT_DST = "TRUE",
  OUTPUT_AOD = "FALSE",
  OUTPUT_WVP = "FALSE",
  OUTPUT_VZN = "TRUE",
  OUTPUT_HOT = "TRUE",
  OUTPUT_OVV = "TRUE"
)
}
\arguments{
\item{OUTPUT_OVV}{}
}
\value{
A data frame containing the parameters and its values
}
\description{
More info at
\url{https://force-eo.readthedocs.io/en/latest/components/lower-level/level2/param.html}
}
