% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_supporting_data.R
\name{getGrid}
\alias{getGrid}
\title{Extract the extent of each grid tile that covers an area of interest}
\usage{
getGrid(l2folder, cubefolder, ext)
}
\arguments{
\item{l2folder}{directory of the level2 data cube}

\item{cubefolder}{folder inside the level2 directory where the data cube is stored}

\item{ext}{extent of the area of interest}
}
\value{
list of extents
}
\description{
Extract the extent of each grid tile that covers an area of interest
}
