% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{getScenes}
\alias{getScenes}
\title{Download Landsat and/or Sentinel-2 data using FORCE}
\usage{
getScenes(
  ext,
  queuefile,
  l1folder,
  metafolder,
  tmpfolder,
  cld = c(0, 100),
  starttime = c(1960, 1, 1),
  endtime = c(),
  tiers = "T1",
  sensors = c("LC08", "LE07", "LT05", "LT04", "S2A", "S2B", "S2A", "S2B")
)
}
\arguments{
\item{ext}{extent of the area of interest, vector with xmin, xmax, ymin, ymax in degrees}

\item{queuefile}{full path to the FORCE queue file}

\item{l1folder}{full path to the FORCE level 1 folder}

\item{metafolder}{full path to the FORCE metadata folder}

\item{tmpfolder}{full path to a temporary folder}

\item{cld}{minimum and maximum cloud cover, e.g. c(0,50)}

\item{starttime}{start date, given as a vector of year, month, and day}

\item{endtime}{end date, given as a vector of year, month, and day}

\item{tiers}{tier of interest (only relevant for Landsat)}

\item{sensors}{vector of sensors, e.g. c('LC08', 'LE07', 'LT05', 'LT04', 'S2A', 'S2B')}
}
\value{
downloads files
}
\description{
Download Landsat and/or Sentinel-2 data using FORCE
}
