% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{import_params}
\alias{import_params}
\title{Import parameters}
\usage{
import_params(file = "data/param/l2param.prm")
}
\arguments{
\item{file}{The parameters file}
}
\value{
The data in the file, line by line, filtering out non-data lines
}
\description{
Checks the FORCE parameter file, and imports its contents as a list of lines
}
