% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landsat.R
\name{Landsat2L2}
\alias{Landsat2L2}
\title{Process Landsat level 1 data to level 2, generate a vrt and summarize the log files}
\usage{
Landsat2L2(
  paramfolder,
  paramfile,
  l2folder,
  LSscenes,
  logfolder,
  Sskiplogfile,
  Ssuccesslogfile,
  Smissionlogfile,
  Sotherlogfile
)
}
\arguments{
\item{paramfolder}{path to the directory with the parameter file}

\item{paramfile}{name of the parameter file}

\item{l2folder}{path to the level 2 folder}

\item{LSscenes}{names of the processed Landsat scenes}

\item{logfolder}{full path to the folder where the logfiles are located}

\item{Sskiplogfile}{full path to the log file for skipped scenes}

\item{Ssuccesslogfile}{full path to the log file for successful scenes}

\item{Smissionlogfile}{full path to the log file for scenes with unknown mission}

\item{Sotherlogfile}{full path to the log file for other scenes}
}
\value{
processes level 1 landsat scenes
}
\description{
Process Landsat level 1 data to level 2, generate a vrt and summarize the log files
}
