% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{loadRData}
\alias{loadRData}
\title{Load RData file and returns it. This function is a substitute for the load function, allowing to assign a user defined variable name when loading a RData file.}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{the path to the Rdata file that needs to be loaded}
}
\value{
R object
}
\description{
Load RData file and returns it. This function is a substitute for the load function, allowing to assign a user defined variable name when loading a RData file.
}
