% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{log1starg}
\alias{log1starg}
\title{Logs the first argument passed to a function with minimal effort}
\usage{
log1starg(f, filename = "log1starg.txt")
}
\arguments{
\item{f}{The function to be called.}

\item{filename}{The log file (default = 'log1starg.txt').}
}
\value{
Void. The argument is written to the desired file.
}
\description{
This function will help us rewriting the system calls in a better way.
}
\examples{
\dontrun{
Imagine you want to log the first argument of the call:
paste("Hi", "there!") # i.e.: you want to log "Hi"
you can do it by:
log1starg(paste)("Hi", "there")

This will be particularly useful in the structures:
system(paste0("text", par, "text", par, ...), intern = TRUE, ignore.stderr = TRUE)
By substituting system by log1starg(system) all the console commands will be
logged in a human readable format.
}
}
