% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_supporting_data.R
\name{makeMaskNoFire}
\alias{makeMaskNoFire}
\title{Make a mask without fire information
The mask identifies pixels that (i) did not have a non-natural land cover type,
(ii) have a tree cover higher than a user-defined threshold and
(iii) were classified as forest in a user-defined year}
\usage{
makeMaskNoFire(lc, lcDates, han, extfolder, Tyr, Ttree)
}
\arguments{
\item{lc}{land cover raster stack (SpatRaster object; terra package)}

\item{lcDates}{dates associated with the layers of the land cover raster stack}

\item{han}{Hansen tree cover raster (SpatRaster object; terra package)}

\item{extfolder}{directory where temporary files should be stored}

\item{Tyr}{year that pixel should be forested}

\item{Ttree}{threshold on tree cover percentage}
}
\value{
SpatRaster layer with data mask (1 equals )
}
\description{
Make a mask without fire information
The mask identifies pixels that (i) did not have a non-natural land cover type,
(ii) have a tree cover higher than a user-defined threshold and
(iii) were classified as forest in a user-defined year
}
