% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{max_narm}
\alias{max_narm}
\title{Maximum value without NA value}
\usage{
max_narm(x, ...)
}
\arguments{
\item{x}{vector of observations}

\item{...}{other options}
}
\value{
maximum value
}
\description{
Maximum value without NA value
}
