% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_all}
\alias{parse_all}
\title{Import parameters as data frame}
\usage{
parse_all(file = "data/param/l2param.prm")
}
\arguments{
\item{file}{The parameters file}
}
\value{
The contents of the file, as a data frame
}
\description{
Checks the FORCE parameter file, and imports its contents as a data frame
}
