% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_keyvalue}
\alias{parse_keyvalue}
\title{Parse a key-value pair from the FORCE parameters file}
\usage{
parse_keyvalue(key, file = "data/param/l2param.prm")
}
\arguments{
\item{key}{Search query (use regex)}

\item{file}{The parameters file}
}
\value{
The value
}
\description{
Parse a key-value pair from the FORCE parameters file
}
