% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_params}
\alias{parse_params}
\title{Parse FORCE parameters file}
\usage{
parse_params(file = "data/param/l2param.prm", key = "")
}
\arguments{
\item{file}{The parameters file}

\item{key}{Search key. Leave empty for parsing the whole file}
}
\value{
The contents of the parameters file as a data frame
}
\description{
More info at \url{https://force-eo.readthedocs.io/en/latest/components/lower-level/level2/param.html}
}
