% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetools.R
\name{partition_dates}
\alias{partition_dates}
\title{Given a start and an endtime, creates n partitions of equal or similar sizes}
\usage{
partition_dates(starttime, endtime, n)
}
\arguments{
\item{starttime}{Start time, as a vector}

\item{endtime}{End time, as a vector}

\item{n}{Number of desired partitions}
}
\value{
A vector of intervals containing all partitions
}
\description{
Consider using: lapply(int_start(i), date_to_vec)
}
