% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_supporting_data.R
\name{prepFire}
\alias{prepFire}
\title{Prepare CCI fire data
Create a raster stack of the CCI fire data, crop the stack to the extent of interest, select the period of interest, and convert to binary rasters (1 = fire, 0 = no fire).}
\usage{
prepFire(
  fcl,
  fjd,
  fdts,
  han,
  msk,
  tempRes,
  Tconf,
  starttime,
  endtime,
  extfolder
)
}
\arguments{
\item{fcl}{a SpatRaster stack with the fire confidence layers}

\item{fjd}{a SpatRaster stack with the fire julian day of year layers}

\item{fdts}{Date object associated with fire layers}

\item{han}{SpatRaster layer with output grid of interest}

\item{msk}{a mask indicating which pixels should be stored (0=should not be processed, 1=should be processed), should have the same grid as the han layer}

\item{tempRes}{temporal resolution of interest for the output fire SpatRaster stack, can be 'monthly', 'daily', or 'quart'}

\item{Tconf}{threshold on the fire confidence, only observations with a fire confidence higher than the threshold are considered to be a true fire}

\item{starttime}{start time of study period of interest (vector with year, month, day)}

\item{endtime}{end time of study period of interest (vector with year, month, day)}

\item{extfolder}{directory where temporary files are stored}
}
\value{

}
\description{
Prepare CCI fire data
Create a raster stack of the CCI fire data, crop the stack to the extent of interest, select the period of interest, and convert to binary rasters (1 = fire, 0 = no fire).
}
