% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_supporting_data.R
\name{prepTreecover}
\alias{prepTreecover}
\title{Prepare tree cover dataset
Crop the data to the extent of interest, mask out areas of no data and - in case the study area covers multiple rasters - merge rasters}
\usage{
prepTreecover(datafolder, ext, fname = "treecover.tif", hanfiles, hanmaskfiles)
}
\arguments{
\item{datafolder}{folder where the processed tree cover data should be stored}

\item{ext}{geographic extent of interest}

\item{fname}{name of the output file}

\item{hanfiles}{list of the tree cover files}

\item{hanmaskfiles}{list of the tree cover mask files}
}
\value{
writes processed raster to disk
}
\description{
Prepare tree cover dataset
Crop the data to the extent of interest, mask out areas of no data and - in case the study area covers multiple rasters - merge rasters
}
