% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{setFolders}
\alias{setFolders}
\title{Set and generate folder structure to store data for the FORCE processing workflow}
\usage{
setFolders(forcefolder)
}
\arguments{
\item{forcefolder}{the main folder where all data needs to be stored (full path)}
}
\value{
generates a folder structure
}
\description{
Set and generate folder structure to store data for the FORCE processing workflow
}
