% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{toFireTS}
\alias{toFireTS}
\title{Create a time series with the desired temporal resolution that has the value 0 if no fire occurs and the value 1 if a fire occurs}
\usage{
toFireTS(x, dts, resol, thres = 95)
}
\arguments{
\item{x}{a time series, the first n values contain the fire confidence values, the subsequent n values the doy of the fires}

\item{dts}{date values associated with the time series observations (Date object)}

\item{resol}{desired temporal resolution of the output}

\item{thres}{threshold on the fire confidence (value between 0 and 100). The higher the confidence value, the more likely the fire really occured. Only fires with a value above the threshold are taken into account.}
}
\value{
vector with 0 if no fire occurs and 1 if a fire occurs
}
\description{
Create a time series with the desired temporal resolution that has the value 0 if no fire occurs and the value 1 if a fire occurs
}
