% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{toRegularTS}
\alias{toRegularTS}
\title{Create regular time series}
\usage{
toRegularTS(tsi, dts, fun, resol)
}
\arguments{
\item{tsi}{vector of observations}

\item{dts}{dates associated with the observations. This should be a Date object.}

\item{fun}{function used to aggregate observations to monthly observations. This should be 'max' or 'mean'.}

\item{resol}{desired temporal resolution of the output. This could be 'monthly', 'quart', or 'daily'}
}
\value{
a vector with a regular time series object
}
\description{
Create regular time series
}
