% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{toShp}
\alias{toShp}
\title{Generate shapefile over AOI}
\usage{
toShp(ext, ofile)
}
\arguments{
\item{ext}{extent of the area of interest, vector with xmin, xmax, ymin, ymax in degrees}

\item{ofile}{full path to the output file name, file should have an '.shp' extension}
}
\value{
generates a shapefile
}
\description{
Generate shapefile over AOI
}
