% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_supporting_data.R
\name{createFireStack}
\alias{createFireStack}
\title{Generate a binary fire stack (0=no fire, 1=fire) at the temporal resolution of interest}
\usage{
createFireStack(msk, fcl, fjd, dts, resol, thres, extfolder)
}
\arguments{
\item{msk}{mask indicating which pixels should be processed (0=not processed, 1 = processed)}

\item{fcl}{SpatRaster stack with fire confidence layers}

\item{fjd}{SpatRaster stack with fire julian date of year}

\item{dts}{dates associated with the fire layers}

\item{resol}{temporal resolution of the ouput stack, can be 'daily', 'monthly', or 'quart'}

\item{thres}{threshold on the fire confidence, observations with a fire confidence higher than the threshold are considered to be a true fire}

\item{extfolder}{folder where temporary files are stored}
}
\value{

}
\description{
Generate a binary fire stack (0=no fire, 1=fire) at the temporal resolution of interest
}
