% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WVP.R
\name{dllWVP}
\alias{dllWVP}
\title{Download Water Vapor Pressure data}
\usage{
dllWVP(wvpfolder, logfile)
}
\arguments{
\item{wvpfolder}{folder where the data should be stored}

\item{logfile}{log file to keep track of the failed downloads}
}
\value{
stores data in output folder
}
\description{
Download Water Vapor Pressure data
}
