% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{gen_params}
\alias{gen_params}
\title{Auxiliary function for generating a parameters' data frame}
\usage{
gen_params(
  FILE_QUEUE = "data/level1/queue.txt",
  DIR_LEVEL2 = "data/level2",
  DIR_LOG = "data/log",
  DIR_TEMP = "data/temp",
  FILE_DEM = "data/misc/dem/srtm.vrt",
  DEM_NODATA = "-32768",
  DO_REPROJ = "TRUE",
  DO_TILE = "TRUE",
  FILE_TILE = "NULL",
  TILE_SIZE = "3000",
  BLOCK_SIZE = "300",
  RESOLUTION_LANDSAT = "30",
  RESOLUTION_SENTINEL2 = "10",
  ORIGIN_LON = "-90",
  ORIGIN_LAT = "60",
  PROJECTION = "GLANCE7",
  RESAMPLING = "NN",
  DO_ATMO = "TRUE",
  DO_TOPO = "TRUE",
  DO_BRDF = "TRUE",
  ADJACENCY_EFFECT = "TRUE",
  MULTI_SCATTERING = "TRUE",
  DIR_WVPLUT = "data/misc/wvp",
  WATER_VAPOR = "NULL",
  DO_AOD = "TRUE",
  DIR_AOD = "NULL",
  ERASE_CLOUDS = "FALSE",
  MAX_CLOUD_COVER_FRAME = "75",
  MAX_CLOUD_COVER_TILE = "75",
  CLOUD_BUFFER = "300",
  SHADOW_BUFFER = "90",
  SNOW_BUFFER = "30",
  CLOUD_THRESHOLD = "0.225",
  SHADOW_THRESHOLD = "0.02",
  RES_MERGE = "REGRESSION",
  DIR_COREG_BASE = "NULL",
  COREG_BASE_NODATA = "-9999",
  IMPULSE_NOISE = "TRUE",
  BUFFER_NODATA = "FALSE",
  TIER = "1",
  NPROC = "1",
  NTHREAD = "1",
  PARALLEL_READS = "FALSE",
  DELAY = "10",
  TIMEOUT_ZIP = "30",
  OUTPUT_FORMAT = "GTiff",
  OUTPUT_DST = "TRUE",
  OUTPUT_AOD = "FALSE",
  OUTPUT_WVP = "FALSE",
  OUTPUT_VZN = "TRUE",
  OUTPUT_HOT = "TRUE",
  OUTPUT_OVV = "TRUE",
  as.list = FALSE
)
}
\arguments{
\item{FILE_QUEUE}{See link above}

\item{DIR_LEVEL2}{See link above}

\item{DIR_LOG}{See link above}

\item{DIR_TEMP}{See link above}

\item{FILE_DEM}{See link above}

\item{DEM_NODATA}{See link above}

\item{DO_REPROJ}{See link above}

\item{DO_TILE}{See link above}

\item{FILE_TILE}{See link above}

\item{TILE_SIZE}{See link above}

\item{BLOCK_SIZE}{See link above}

\item{RESOLUTION_LANDSAT}{See link above}

\item{RESOLUTION_SENTINEL2}{See link above}

\item{ORIGIN_LON}{See link above}

\item{ORIGIN_LAT}{See link above}

\item{PROJECTION}{See link above}

\item{RESAMPLING}{See link above}

\item{DO_ATMO}{See link above}

\item{DO_TOPO}{See link above}

\item{DO_BRDF}{See link above}

\item{ADJACENCY_EFFECT}{See link above}

\item{MULTI_SCATTERING}{See link above}

\item{DIR_WVPLUT}{See link above}

\item{WATER_VAPOR}{See link above}

\item{DO_AOD}{See link above}

\item{DIR_AOD}{See link above}

\item{ERASE_CLOUDS}{See link above}

\item{MAX_CLOUD_COVER_FRAME}{See link above}

\item{MAX_CLOUD_COVER_TILE}{See link above}

\item{CLOUD_BUFFER}{See link above}

\item{SHADOW_BUFFER}{See link above}

\item{SNOW_BUFFER}{See link above}

\item{CLOUD_THRESHOLD}{See link above}

\item{SHADOW_THRESHOLD}{See link above}

\item{RES_MERGE}{See link above}

\item{DIR_COREG_BASE}{See link above}

\item{COREG_BASE_NODATA}{See link above}

\item{IMPULSE_NOISE}{See link above}

\item{BUFFER_NODATA}{See link above}

\item{TIER}{See link above}

\item{NPROC}{See link above}

\item{NTHREAD}{See link above}

\item{PARALLEL_READS}{See link above}

\item{DELAY}{See link above}

\item{TIMEOUT_ZIP}{See link above}

\item{OUTPUT_FORMAT}{See link above}

\item{OUTPUT_DST}{See link above}

\item{OUTPUT_AOD}{See link above}

\item{OUTPUT_WVP}{See link above}

\item{OUTPUT_VZN}{See link above}

\item{OUTPUT_HOT}{See link above}

\item{OUTPUT_OVV}{See link above}

\item{as.list}{Output a list instead of a data.frame}
}
\value{
A data frame or list containing the parameters and its values
}
\description{
More info at
\url{https://force-eo.readthedocs.io/en/latest/components/lower-level/level2/param.html}
}
