% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{import_params}
\alias{import_params}
\title{Import parameters as data frame}
\usage{
import_params(file = "data/param/l2param.prm")
}
\arguments{
\item{file}{The parameters file}
}
\value{
The contents of the file, as a data frame
}
\description{
Checks the FORCE parameter file, and imports its contents as a data frame
More info at
\url{https://force-eo.readthedocs.io/en/latest/components/lower-level/level2/param.html}
}
