% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_supporting_data.R
\name{prepFire}
\alias{prepFire}
\title{Prepare CCI fire data
Create a raster stack of the CCI fire data, crop the stack to the extent of interest, and select the period of interest.}
\usage{
prepFire(
  ifolder,
  datafolder,
  ext,
  fjdname = "fireJD.tif",
  fclname = "fireCL.tif",
  startyr,
  endyr,
  fireclfiles,
  firejdfiles
)
}
\arguments{
\item{ifolder}{folder where the CCI file files are stored}

\item{datafolder}{folder where the output files should be stored}

\item{ext}{geographic extent of interest}

\item{fjdname}{output file name of the fire julian date raster stack}

\item{fclname}{ouput file name of the fire confidence raster stack}

\item{startyr}{start date of the observation period}

\item{endyr}{end date of the observation period}

\item{fireclfiles}{names of the fire confidence layers}

\item{firejdfiles}{names of the fire julian date layers}
}
\value{
saves raster stacks to disk
}
\description{
Prepare CCI fire data
Create a raster stack of the CCI fire data, crop the stack to the extent of interest, and select the period of interest.
}
