% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_supporting_data.R
\name{prepLandcover}
\alias{prepLandcover}
\title{Prepare the MapBiomas land cover data:
Crop the land cover data to the desired extent and extract the study period of interest}
\usage{
prepLandcover(
  ifolder,
  datafolder,
  ext,
  fname = "landcover.tif",
  startyr,
  endyr
)
}
\arguments{
\item{ifolder}{Directory of the Mapbiomas files}

\item{datafolder}{Directory where the processed land cover data will be stored}

\item{ext}{Geographic extent that should be processed}

\item{fname}{Output file name}

\item{startyr}{start date of study period}

\item{endyr}{end date of study period}
}
\value{

}
\description{
Prepare the MapBiomas land cover data:
Crop the land cover data to the desired extent and extract the study period of interest
}
